backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9

# PDF Parsing
PyMuPDF>=1.24.0

# Embeddings & LLM
openai>=1.40.0
anthropic>=0.34.0
httpx>=0.27.0

# Vector Stores
faiss-cpu>=1.8.0
numpy>=1.26.0
chromadb>=0.5.0

# Optional: HuggingFace embeddings (uncomment if using huggingface provider)
# sentence-transformers>=3.0.0
